pyproject.toml
[tool.poetry]
name = "deposit-app"
version = "0.1.0"
description = ""
authors = ["Lex <alexnosov2009@rambler.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
